field via relationship with type, alias (nil wild) calculation
